+++ NEWS.md
@@ -1,6 +1,8 @@
-# CDSE 0.2.1.9000
+# CDSE 0.3.O (2025-08-05)
 
 -   Adapted `GetCollections` to deal with multi-instrument collections like 'Sentinel 3 Synergy L2'
+-   Added `MakeEvalScript` function that generates evalscript from the spectral indices defined in the package `rsi`
+-   Expanded documentation on evalscripts
 -   Fixed internal function `CheckBbox`
 
 # CDSE 0.2.1 (2024-08-16)